requirements.txt
streamlit>=1.28
pandas>=2.0
numpy>=1.24
Pillow>=9.5
tensorflow>=2.12
